[ads][CodeHealth] IWYU, size_t, std::string_view, spelling (#32283)
IWYU. Use `size_t` and `std::string_view` where
appropriate. Fix spellings. Remove ::testing::Invoke.

+++ components/brave_ads/core/internal/user_engagement/ad_events/search_result_ads/search_result_ad_event_handler_util_for_non_rewards_unittest.cc
@@ -3,6 +3,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
+#include <cstddef>
+
 #include "base/run_loop.h"
 #include "base/test/gmock_callback_support.h"
 #include "base/test/mock_callback.h"
@@ -70,8 +72,9 @@ TEST_F(BraveAdsSearchResultAdEventHandlerUtilForNonRewardsTest,
           /*should_generate_random_uuids=*/true);
 
   // Act
-  for (int i = 0;
-       i < static_cast<int>(mojom::SearchResultAdEventType::kMaxValue); ++i) {
+  for (size_t i = 0;
+       i < static_cast<size_t>(mojom::SearchResultAdEventType::kMaxValue);
+       ++i) {
     const auto mojom_ad_event_type =
         static_cast<mojom::SearchResultAdEventType>(i);
     if (mojom_ad_event_type != mojom::SearchResultAdEventType::kClicked) {
@@ -141,8 +144,9 @@ TEST_F(BraveAdsSearchResultAdEventHandlerUtilForNonRewardsTest,
           /*should_generate_random_uuids=*/true);
 
   // Act & Assert
-  for (int i = 0;
-       i < static_cast<int>(mojom::SearchResultAdEventType::kMaxValue); ++i) {
+  for (size_t i = 0;
+       i < static_cast<size_t>(mojom::SearchResultAdEventType::kMaxValue);
+       ++i) {
     const auto mojom_ad_event_type =
         static_cast<mojom::SearchResultAdEventType>(i);
     if (mojom_ad_event_type != mojom::SearchResultAdEventType::kClicked) {
