+++ components/brave_ads/core/internal/user_engagement/ad_events/search_result_ads/search_result_ad_event_handler_util_for_rewards_unittest.cc
@@ -3,6 +3,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
+#include <cstddef>
+
 #include "base/run_loop.h"
 #include "base/test/gmock_callback_support.h"
 #include "base/test/mock_callback.h"
@@ -64,8 +66,9 @@ TEST_F(BraveAdsSearchResultAdEventHandlerUtilForRewardsTest,
           /*should_generate_random_uuids=*/true);
 
   // Act
-  for (int i = 0;
-       i < static_cast<int>(mojom::SearchResultAdEventType::kMaxValue); ++i) {
+  for (size_t i = 0;
+       i < static_cast<size_t>(mojom::SearchResultAdEventType::kMaxValue);
+       ++i) {
     const auto mojom_ad_event_type =
         static_cast<mojom::SearchResultAdEventType>(i);
     if (mojom_ad_event_type !=
@@ -113,8 +116,9 @@ TEST_F(BraveAdsSearchResultAdEventHandlerUtilForRewardsTest,
           /*should_generate_random_uuids=*/true);
 
   // Act & Assert
-  for (int i = 0;
-       i < static_cast<int>(mojom::SearchResultAdEventType::kMaxValue); ++i) {
+  for (size_t i = 0;
+       i < static_cast<size_t>(mojom::SearchResultAdEventType::kMaxValue);
+       ++i) {
     const auto mojom_ad_event_type =
         static_cast<mojom::SearchResultAdEventType>(i);
     EXPECT_TRUE(IsAllowedToFireAdEvent(mojom_creative_ad, mojom_ad_event_type));
@@ -129,8 +133,9 @@ TEST_F(BraveAdsSearchResultAdEventHandlerUtilForRewardsTest,
           /*should_generate_random_uuids=*/true);
 
   // Act & Assert
-  for (int i = 0;
-       i < static_cast<int>(mojom::SearchResultAdEventType::kMaxValue); ++i) {
+  for (size_t i = 0;
+       i < static_cast<size_t>(mojom::SearchResultAdEventType::kMaxValue);
+       ++i) {
     const auto mojom_ad_event_type =
         static_cast<mojom::SearchResultAdEventType>(i);
     EXPECT_TRUE(IsAllowedToFireAdEvent(mojom_creative_ad, mojom_ad_event_type));
