+++ components/brave_ads/core/internal/user_engagement/conversions/conversions_util_internal_unittest.cc
@@ -5,6 +5,8 @@
 
 #include "brave/components/brave_ads/core/internal/user_engagement/conversions/conversions_util_internal.h"
 
+#include <cstddef>
+
 #include "brave/components/brave_ads/core/internal/ad_units/ad_test_util.h"
 #include "brave/components/brave_ads/core/internal/common/test/test_base.h"
 #include "brave/components/brave_ads/core/internal/common/test/time_test_util.h"
@@ -26,14 +28,14 @@ TEST_F(BraveAdsConversionsUtilInternalTest,
   test::DisableBraveRewards();
 
   // Act & Assert
-  for (int i = 0; i < static_cast<int>(mojom::AdType::kMaxValue); ++i) {
+  for (size_t i = 0; i < static_cast<size_t>(mojom::AdType::kMaxValue); ++i) {
     const auto mojom_ad_type = static_cast<mojom::AdType>(i);
 
     const AdInfo ad = test::BuildAd(mojom_ad_type,
                                     /*should_generate_random_uuids=*/false);
 
-    for (int j = 0; j < static_cast<int>(mojom::ConfirmationType::kMaxValue);
-         ++j) {
+    for (size_t j = 0;
+         j < static_cast<size_t>(mojom::ConfirmationType::kMaxValue); ++j) {
       const auto confirmation_type = static_cast<mojom::ConfirmationType>(j);
 
       bool expected_can_convert_ad_event;
@@ -59,14 +61,14 @@ TEST_F(BraveAdsConversionsUtilInternalTest,
 
 TEST_F(BraveAdsConversionsUtilInternalTest, CanConvertAdEventForRewardsUser) {
   // Act & Assert
-  for (int i = 0; i < static_cast<int>(mojom::AdType::kMaxValue); ++i) {
+  for (size_t i = 0; i < static_cast<size_t>(mojom::AdType::kMaxValue); ++i) {
     const auto mojom_ad_type = static_cast<mojom::AdType>(i);
 
     const AdInfo ad = test::BuildAd(mojom_ad_type,
                                     /*should_generate_random_uuids=*/false);
 
-    for (int j = 0; j < static_cast<int>(mojom::ConfirmationType::kMaxValue);
-         ++j) {
+    for (size_t j = 0;
+         j < static_cast<size_t>(mojom::ConfirmationType::kMaxValue); ++j) {
       const auto confirmation_type = static_cast<mojom::ConfirmationType>(j);
 
       // Only viewed and clicked ad events are allowed to be converted.
